backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
httpx==0.27.0
supabase==2.9.0
openai==1.50.0
stripe==10.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
websockets==12.0
pytest==8.3.0
pytest-asyncio==0.24.0
